Less annoying concept mouse-out behaviour
I'm trying to smooth out an annoying interaction in trying to rename a
concept. Before this commit, I would lose focus in the middle of
selecting the concept label because I would unconsciously click-n-drag
as though I were highlighting text, and in doing so would often bump
the mouse out of focus, which disables everything.

Now, once you click the text label, we ignore the mouse-out until you've
typed at least a character in the box