Stop tracking context effects if they are unset
There's a bit of trickiness here. This can cause a subtle bug
if you rely on membership in the context effects dictionary to
determine if an object can be painted. Suppose for example you
had a curve that was highlighted green, and you want to remove
the effect (which consists of changing the colour to black).
Because we are modeling this as removing an effect, the curve
is deleted from the dictionary, and so we skip the curve when
we want to apply the black paint... which results in the curve
seeming to be permanently highlighted