Tighten up key equality in effects system
This is partly just avoiding Lombok EqualsAndHashcode unless we really
mean to use it. Nothing particularly wrong with it, mind you, just a
matter of tightening things up.

Also: we want to base our notion of key equality on reference equality
of the objects; because here the intention of the key is to refer to the
object itself, not its contents/state. Without this fix, the effects
tracking system can fail to recognise it's talking about the same key as
the the underlying object is mutated. It's possible that the
Lombok-generated version of equals and hashcode was doing the right
thing here, but I don't understand why it should